requirements.txt
speechrecognition
pyttsx3
scikit-learn
tensorflow
pygame
pyautogui
numpy
pandas
PyAudio
google-generativeai
opencv-python
Pillow
mss
screeninfo
dxcam
pytesseract
